move recreate cmd tests to own module
